Implement friend notifier
This will keep track of friends presence and state, using the chrome.alarms and chrome.storage.session APIs.

On each execution, will load the authenticated users friends, and their presences. If a presence changes, and we deem it worthy of a notification, we will send a notification.

If the authenticated user is not following their friend, we will ignore notifications for this friend.

TODO: Allow users to ignore specific types of presence changes.